+++ go/understack/.goreleaser.yaml
@@ -18,7 +18,7 @@ builds:
       - CGO_ENABLED=0
     goos: ["linux", "darwin", "windows"]
     goarch: ["386", "amd64", "arm64"]
-    binary: deploy-cli
+    binary: understack
     ldflags:
       - -s -w -X "main.version={{.Version}}"
