+++ go/understack/cmd/argocd/secrets.go
@@ -6,8 +6,8 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/rackerlabs/understack/go/deploy-cli/cmd"
-	"github.com/rackerlabs/understack/go/deploy-cli/helpers"
+	"github.com/rackerlabs/understack/go/understack/cmd"
+	"github.com/rackerlabs/understack/go/understack/helpers"
 
 	"github.com/charmbracelet/log"
 	"github.com/gookit/goutil/envutil"
