+++ go/understack/cmd/argocd/templates/argoCluster.tmpl
