+++ go/understack/cmd/certManager/secrets.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/rackerlabs/understack/go/deploy-cli/cmd"
-	"github.com/rackerlabs/understack/go/deploy-cli/helpers"
+	"github.com/rackerlabs/understack/go/understack/cmd"
+	"github.com/rackerlabs/understack/go/understack/helpers"
 
 	"github.com/charmbracelet/log"
 	"github.com/gookit/goutil/envutil"
