+++ go/understack/cmd/init/init.go
@@ -5,14 +5,14 @@ import (
 	"os"
 	"os/exec"
 
-	"github.com/rackerlabs/understack/go/deploy-cli/cmd"
-	"github.com/rackerlabs/understack/go/deploy-cli/cmd/argocd"
-	"github.com/rackerlabs/understack/go/deploy-cli/cmd/certManager"
-	"github.com/rackerlabs/understack/go/deploy-cli/cmd/dex"
-	"github.com/rackerlabs/understack/go/deploy-cli/cmd/helmConfig"
-	"github.com/rackerlabs/understack/go/deploy-cli/cmd/node"
-	"github.com/rackerlabs/understack/go/deploy-cli/cmd/openstack"
-	"github.com/rackerlabs/understack/go/deploy-cli/cmd/other"
+	"github.com/rackerlabs/understack/go/understack/cmd"
+	"github.com/rackerlabs/understack/go/understack/cmd/argocd"
+	"github.com/rackerlabs/understack/go/understack/cmd/certManager"
+	"github.com/rackerlabs/understack/go/understack/cmd/dex"
+	"github.com/rackerlabs/understack/go/understack/cmd/helmConfig"
+	"github.com/rackerlabs/understack/go/understack/cmd/node"
+	"github.com/rackerlabs/understack/go/understack/cmd/openstack"
+	"github.com/rackerlabs/understack/go/understack/cmd/other"
 
 	"github.com/charmbracelet/log"
 	"github.com/gookit/goutil/envutil"
