+++ go/understack/cmd/node/node.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"os"
 
-	"github.com/rackerlabs/understack/go/deploy-cli/cmd"
-	"github.com/rackerlabs/understack/go/deploy-cli/helpers"
+	"github.com/rackerlabs/understack/go/understack/cmd"
+	"github.com/rackerlabs/understack/go/understack/helpers"
 
 	"github.com/charmbracelet/log"
