+++ go/understack/helpers/random.go
